Add files via upload
This project focuses on building a machine learning model to classify emails as spam or not spam (ham). The workflow includes text preprocessing, feature extraction, and the implementation of classification algorithms to accurately detect spam emails. The goal is to create an efficient email filtering system that can be applied in real-world scenarios for spam detection.
*Key Skills Learned:
-Text preprocessing (cleaning, tokenization, stemming/lemmatization)
-Feature extraction (TF-IDF, Bag of Words)
-Implementation of classification algorithms (e.g., Naive Bayes, Logistic Regression, SVM)
*Real-World Application:
This model can be integrated into email systems to automatically filter out spam, improving user experience and security.